+++ test/test_variogram.py
@@ -16,7 +16,7 @@ GRID = grid_from_resolution(
 
 DIST = grid_to_distance_matrix(
     GRID, lat_coord="latitude", lon_coord="longitude"
-).values
+)
 
 
 @pytest.mark.parametrize(
@@ -113,7 +113,13 @@ DIST = grid_to_distance_matrix(
 def test_variogram(variogram_model, parameters, variance):
     # TEST: Generates a valid covariance
     variogram = variogram_model(**parameters)
-    variogram_result = variogram.fit(DIST)
+    variogram_result_xr = variogram.fit(DIST)
+    variogram_result = variogram.fit(DIST.values)
+
+    # TEST: xarray and numpy versions are the same
+    assert np.allclose(variogram_result, variogram_result_xr.values)
+    variogram_result_xr = None
+
     covariance = variogram_to_covariance(variogram_result, variance)
 
     evals = np.linalg.eigvalsh(covariance)